+++ src/Backport.jl
@@ -24,6 +24,7 @@ macro backport()
     expr = Expr(:block)
     #VERSION < v"1.2.0-rc1" && push!(expr, :(const mapreduce = Backport.mapreduce))
     VERSION < v"1.2.0-rc1" && push!(expr.args, :(using Backport: mapreduce))
+    VERSION < v"1.6.0-beta1" && push!(expr.args, :(using Backport: reverse, reverse!))
     esc(expr)
 end
 
@@ -62,6 +63,37 @@ if VERSION < v"1.7.0-beta1"
     end
     (obj::Returns)(@nospecialize(args...); @nospecialize(kwds...)) = getfield(obj, :value)
     eval(:(export Returns))
+end
+
+if VERSION < v"1.6.0-beta1"
+    # Prior to Julia 1.6, `reverse` for all dimensions is not defined and `reverse!` does
+    # not accept keywords. NOTE Helper functions `_reverse` and `_reverse!` are introduced
+    # for inference to work.
+    reverse(args...; kwds...) = Base.reverse(args...; kwds...)
+    reverse!(args...; kwds...) = Base.reverse!(args...; kwds...)
+    reverse(A::AbstractArray; dims = :) = _reverse(A, dims)
+    reverse!(A::AbstractArray; dims = :) = _reverse!(A, dims)
+    _reverse(A::AbstractVector, d::Integer) =
+        isone(d) ? Base.reverse(A) : throw(ArgumentError("invalid dimension $d ≠ 1"))
+    _reverse!(A::AbstractVector, d::Integer) =
+        isone(d) ? Base.reverse!(A) : throw(ArgumentError("invalid dimension $d ≠ 1"))
+    _reverse(A::AbstractArray, d) = Base.reverse(A; dims=d)
+    _reverse!(A::AbstractArray, d) = copyto!(A, Base.reverse(A; dims=d))
+    _reverse(A::AbstractVector, ::Colon) = Base.reverse(A)
+    _reverse!(A::AbstractVector, ::Colon) = Base.reverse!(A)
+    _reverse(A::AbstractArray, ::Colon) = _reverse!(Base.copymutable(A), :)
+    function _reverse!(A::AbstractArray, ::Colon)
+        I = eachindex(A)
+        k = last(I) + first(I)
+        @inbounds for i in I
+            (j = k - i) > i || break
+            Ai = A[i]
+            Aj = A[j]
+            A[i] = Aj
+            A[j] = Ai
+        end
+        return A
+    end
 end
 
 """